agent/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
web3>=6.0.0
httpx>=0.27.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
